+++ src/components/Keyboard/Key/consts.ts
@@ -4,3 +4,6 @@ export const DIFF_DEPTH_BOTTOM = 0.4 / 2
 export const TOP_WIDTH_INSET = 0.07
 export const TOP_DEPTH_INSET = 0.05
 export const HEIGHT = 0.5
+
+export const KEY_DOWN_POSITION = -0.5
+export const KEY_DOWN_DURATION = 0.02
